Initial conduit URI code
This commit contains basic code for handling of conduit URIs and for
describing an actual conduit.  Conduit URIs have the concept of canonical
vs. non-canonical, built around whether external resources must be queried
to turn the conduit URI into a hostname and port; a URI.Canonicalize()
method exists to turn a non-canonical conduit URI into a list of canonical
URIs.