Create and use a custom Conn
This commit creates a custom Conn interface with a GetInfo method.  It
also creates the Info type and updates the Conduit type to embed the Info
instead of using directly specified types.  Finally, a WrapConn type is
created to allow net.Conn to be wrapped up together with an Info.