Simplify Conduit
This commit simplifies the conduit package by creating a single Conduit
type to be returned by mechanisms.  It also unifies the Transport and
Security mechanisms into a Mechanism type and adds Dial and Listen methods
to the URI type, plus top-level Dial and Listen functions that take a URI
string.  This commit does not include any mechanisms.