Specify the config
This commit adds a Config interface that describes what the configuration
must look like.  This allows specific mechanisms to specify their own
configuration data, but still allows a common config type to be passed to
the Dial and Listen functions, which don't yet know what mechanism(s) are
being used.