Add LocalAddr
This commit adds the tooling necessary to make a LocalAddr option work.
Built-in mechanisms using mkDialer or mkListenConfig can additionally pass
an appropriate filter, which is passed the option, and which may return an
error.  The LocalAddr option then works by storing a passed-in URI; the
TCP mechanism can then use TCP-specific tooling to convert from the URI to
the appropriate TCPAddr for use with the LocalAddr field of the Dialer.