Begin the protocol implementation
This commit begins the implementation of the protocol parser and generator
by creating Header and ExtHeader types to represent the carrier protocol
header and the associated extension header.  Methods are provided to fill
in the types from a sequence of bytes, and to fill in a byte buffer with
the encoded forms of the types.